fn_800E1CA8 returns u8 at its definition (a u8 local; exact) and fn_800DCB10 returns pfn1F8's u8 answer (GameEffects tests it; asm 800DCB28 leaves r3 as is): both declared once in game.h (audit C3, C4)

+++ src/GameRound.c
@@ -541,8 +541,8 @@ u8 fn_800E1BBC(void) {
 }
 
 // Whether the current hole is the round's last.
-int fn_800E1CA8(void) {
-    int b = 1;
+u8 fn_800E1CA8(void) {
+    u8 b = 1;
     int i;
     for (i = gpGame->nCurHole + 1; i < 18; i++) {
         if (gpGame->bHoleSelected[i]) {
